// src/models/bop500.ts
// using by src/mocks/subCategoryData and SubCategoryTable but the table is not being used currently
export interface SUBCATEGORY_DATA_TYPE {
  id: string;
  subcategoryName: string;
  allSubcategoryName: string;
  title: string;
  address: string;
  city: string;
  rank: number;
  ratingCount: number;
  rating: number;
  temporaryClosed: number;
}